implement items controller

+++ src/main/java/com/mirera/merchandise/application/port/inbound/items/request/ItemReqDTO.java
@@ -1,11 +1,12 @@
 package com.mirera.merchandise.application.port.inbound.items.request;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public record ItemReqDTO(
   String itemName,
   String itemDescription,
-  double itemPrice,
+  BigDecimal itemPrice,
   String itemImageUrl,
   int stockQuantity,
   String itemColor,
